[PRMDR-167] Rename folder pds_data to test_data, add test data for ods api response
